requirements.txt
fastapi
uvicorn
google-genai
sentence-transformers
faiss-cpu
pydantic
pydantic-settings
python-dotenv
tiktoken
